+++ src/chip8/Display.java
@@ -1,8 +1,8 @@
 package Chip8;
 
 public interface Display {
-	public byte getScreen(byte index);
-	public void setScreen(byte index, byte value);
+	public byte getScreen(int index);
+	public void setScreen(int index, byte value);
 	public void print();
 	public void clear();
 }
